Well, here goes my first commit to LiVid's CVS....
	-Added display_gatos.c to mpeg2dec.
	-Added autoconf updates to mpeg2dec to detect GATOS.

	The display_gatos thingy really doesn't work well yet for any of
the Rage128s.  For any of these cards you have to modify the code (as I
mentioned in my originaly message to livid-dev).  Eventually it'll be nice
and generic, but before I can do that, I have to get it working better on
the 128s.
	I'll put up a change to the autoconf stuff to add a configure
option to disable gatos detection for these cards tomorrow (software is
currently a lot better than using the 128's BES).  I don't know autoconf
very well, the autoconf changes are/will be written by Daniel
Caujolle-Bert, another member of the GATOS team.

	Hopefully I didn't break anything.
						--Insomnia (Stea Greene)