Quick hack to fix bistream so dent will stop complaining about mpeg2dec not working directly in oms :)
bistream.c has been changed to read chunks of data into a 64k buffer
and hence get the required 32 bit align, and multiple of 4 bytes.

And it doesn't even go slower (this is a case of give a little,
take a little). It could probably be faster if I restructured
the bitstream internals a bit.

aaron