Major core reworkage of mpeg2dec to allow libmpeg2 to be used by oms.
The big change is that the mpeg2_decode_frame changed to mpeg2_decode_data
and the fill_buffer callback was removed. mpeg2_decode_data now
returns the number of frames finished during the call. This required a number
of internal changes, and a lot of code was touched.

Unfortunately this change lost us %5 in performance on my machine. Don't
worry though, the optimizations we have planned in the next few weeks should
make this seem like a drop in the bucket.

Incidentally, I have been playing around the perf counter patches on my
celeron 366. I have measured our CPU utilization at 1.5 CPI for a particular
test clip. This is pretty terrible for a OOE cpu and just goes to show
how hamstrung we are by our cache footprint.

cheers,
aaron