Yes, I'm still alive.
This patch ports the X11 code, including Xv, to the new plugin
architecture.  It's no better that it was, but it works with the new
scheme and I also added the framework for me to (finally) add DGA support.

If you don't have shared Xv or Xf86dga libs, you should us "--disable-xv"
and/or "--disable-dga" respectively.

Also cleaned up a little auto* stuff in the video plugins dir.

DGA, faster XV, better X code, and some cool features soon to come....

---Insomnia (Stea Greene)