Rewrote the routine in decode.c that scans for start codes. The new one a bit simpler (not very much though) and it puts the first byte of each chunk (which indicates the chunk type) separately from the others, which I prefer.
A side effect of this is that the data after this chunk type byte is now
aligned, which should fix the problem that hh reported on sparc.
hh, can you confirm me that the problem is gone ?

I was surprised that small hacks in the scancode search can have an impact
on the decoder speed. Depending on how I wrote my small loop I could measure
differences of about 3% on the total decoding + yuv speed ! Maybe it would
be worth trying to unroll this loop, or using a different search algorithm ???? I have not experimented with this, but if someone wants to try, its in
mpeg2_decode_data.

With the fixed alignment and the improved startcode search, I get a speed
improvement of around 2%.

I also included one #ifdef in the X11 video output to allow me to benchmark
things (heh, I was doing that mod in my tree all the time)

walken