rewrite of the slice decoder.
I think it should be easier to implement new motion comp modes using this
scheme. look at motion_frame and motion_field - other modes should be
equally simple to implement.

Fixed a problem with the address increment - if this is the first macroblock
in the slice, then it does not indicate skipped macroblocks, but only the
x offset of the first macroblock in this slice. The problem didnt show up
with dvd streams because they use only one slice header per horizontal line
of macroblocks.

Also got rid of the way mba was used to calculate X and Y positions -
actually mba is now only used to find out when we finish a picture, but
this would be better done in decode.c so mba is on the way out.

MC is now done before IDCT, and the IDCT functions will soon be rewritten
to take advantage of that. basically idct_add or idct_copy could be done
inside of the idct_block function to avoid some unneeded load/store
operations. the slice_non_intra_DCT function could also probably be
optimize to take advantage of the cases where there are very few non-NULL
DCT coefficients - I think the case where there is only a DC coefficient
is probably quite common, but I will need to investigate this.

walken