slice parsing updates :
* modified the bitstream code. Now needbits() is not called by getbits(),
we have to call it explicitely. After we call it we can be sure there is
at least 17 bits in the bit buffer, so we can get or show them without
additional calls to needbits.

* modified the DCT huffman parsing routines, trying to avoid having too many
mispredicted branches in there

This results in a nice two-digit speed improvement for me :)

I may try some additional updates in the DCT parsing later to make it more
robust against errors while keeping it fast :)

walken