Further updates to the DCT parsing code. The code looks funny now but if you only look at the fast path its not so hard to understand and it has way fewer conditionals than before. On my favourite test stream that I always use for reference, I'm now up to 101.5 to 102 fps :)
Yay, we broke the 100 fps barrier :)

This code does not display anything on errors, but its should never
buffer overflow either. My plan is to only check for errors in one
single place - probably get_mba_increment or something - and exit
the slice code from there if needed. Who cares if we dont detect
all errors - if we get garbage streams we can display garbage... the
important thing is that we never segfault on these streams.