finished the huffman parsing rewrite.
bitstream.h and the huffman tables moved into a new file vlc.h
the huffman parsing functions are now in slice.c (because they are inlined)

The error checking in the parsing functions is gone. Who needs to parse
bogus streams anyway :) I just made sure that we do not segfault or go into
infinite loops because of them. There are still two known causes of segfaults
though : the reference is not checked when we do MC, and the destination
is not checked either so it can go over the right edge of the screen. Once
I fix those the decoder should not segfault on bogus streams anymore.

walken