build fixes for libc5 & glibc2.0 systems, which do not have inttypes.h
The default inttypes.h in the compatibility directory will only be used
if inttypes.h is not found in the system headers. It should work for most
people, but if it does not, though for them. I have included compile time
checks so that it breaks if the type sizes are not what we expect.

walken