Moved DCTblock to the slice_t structure.
This is a nice step into making libmpeg2 reentrant - right now I think
the only missing points are motion_dmv_top_field_first and current_field ?
Which means that we are actually reentrant for streams that dont use
field pictures or dmv motion compensation. i.e. 90% of the streams :)

Martin, can you test if that level of reentrancy actually works ?

The slice_t structure is on the stack, let's hope it stays aligned enough...
(it seems to stay aligned, I'm not sure why though).