First part of MC code rework
At some point we need to have the coordinates (in half-pixels)
of the MC source data, so that we can check that we do not copy
from outside of the source picture boundaries. In the previous code,
these coordinates were never explicitly exposed, which made the
checking difficult.

The new MC code still needs a lot of work, but this is a first step.