mlib changes:
use IEEE compliant IDCT routines by default. This avoids error accumulation.

problems:
* this is slower than the non-IEEE mlib IDCT routine
* this is only available for non-intra coded blocks. intra blocks are still
	decoded using a noncompliant IDCT routine. However this is not as
	much of a problem as intra blocks do not cause error accumulation.

The old, faster behaviour can be restored by setting the environment variable
MLIB_NON_IEEE=1

Sometime in the future it would be worth it to come up with a fast
compliant IDCT routine using the VIS instructions - but this can not
be done in the very short term.