get rid of memset, which was used to clear IDCT blocks before we decode the IDCT coefficients.
Now the clearing is done in the IDCT function, after the block has
been copied or added to the target buffer. This allows us to use MMX
or altivec optimizations to do the clearing, and also to interleave it
with other instructions to hide the memory latency.

I figured out this function was slower than I thought, and doing this
change brings about a 5% speed improvement.

I broke the altivec optimizations, I'll have to fix it as soon as I
get my hands on a g4 shell again.