Make sure the MMX/MMXEXT code can go thru the DC-only IDCT fast path just a little more often.
Some background here. Because of mpeg2 oddification, when there is
only one DC coefficient (block[0]) transmitted, block[63] will still
be 1 if block[0] is even. So we have to make sure that even in these
cases, we end up producing the same result as the full IDCT code
would.

In a full-precision IDCT, the DC-only code is equivalent to the full
IDCT whenever (block[0] & 7 != 4) - odd values are no problem since
block[63] is zero then, and the values 0, 2, 6 are OK too because the
rounding happens to work fine then. This also applies to the C IDCT
code as it is precise enough too.

Because the MMX code is not just as precise, it yelds a different
result than the full-precision one when block[0] & 7 == 6. So I've
been preserving this behaviour in my previous code by letting us go
thru the slow path in that case. However the fast path is actually
more precise, so it's OK to go there if I just update the test suite
accordingly.

We still go thru the slow path when (block[0] & 7) == 4, as it's
actually more precise for that case.