This fixes one possible error when dealing with corrupted streams
When parsing a corrupt sequence header, we return STATE_INVALID and
skip to the next sequence, gop or picture header, however we also need
to make sure we do not corrupt the decoder_t state. This fix makes
sure the decoder_t is only updated later on when we finalize the
(correctly received) sequence.