Make mpeg2_seek_header a static function, rename into seek_header
In seek_header, if we receive a picture start code, go to
mpeg2_header_picture_start so we actually go into STATE_PICTURE

When receiving an unexpected startcode after decoding a slice, send a
STATE_INVALID, then try to recover with the right action if this is a
sequence, gop or end_of_sequence header, otherwise (usualy badly muxed
streams) skip to the next startcode and if it's a slice, decode it as
part of the current picture.

All these changes help a bit with badly corrupted and/or badly muxed
streams.