Aspect ratio guessing.
This adds a new function mpeg2_guess_aspect, which tries to figure out
a correct aspect ratio for common video formats derived from BT.601

The following resolutions are detected and assumed to be derived from
BT.601 with the usual 13.5 MHz sampling (59:54 or 10:11 pixel aspect
ratio depending on 525/625 line variant):

In mpeg1 streams:

720x576 (D1, DV, DVB, DVD)
704x576 (1/1 D1, DVB, DVD, 4CIF)
352x576 (D2, 1/2 D1, CVD, DVB, DVD)
352x288 (D4, VCD, DVB, DVD, CIF)
176x144 (QCIF)
720x486 (D1)
704x486
720x480 (DV, DSS, DVD)
704x480 (1/1 D1, ATSC, DVD)
352x480 (D2, 1/2 D1, CVD, DVD)
352x240 (D4, VCD, DSS, DVD)

Additionally, in mpeg2 streams the following resolutions are detected
and assumed to be derived from video signals with 10.125 MHz (3/4) or
9 MHz (2/3) sampling:

544x576 (DVB, laserdisc)
528x576 (3/4 D1, DVB, laserdisc)
480x576 (2/3 D1, DVB, SVCD)
544x480 (DSS, laserdisc)
528x480 (3/4 D1, laserdisc)
480x480 (2/3 D1, SVCD)

These formats are recognised for a display aspect ratio of either 4:3
or 16:9. Additionally, for mpeg1 streams we check that the encoded SAR
is the expected one, and for mpeg2 streams we check that the display
size is the same as the picture size.

See http://www.mir.com/DMG/aspect.html for more information.